Make rating float

+++ database/migrations/2016_12_07_091351_create_rates_table.php
@@ -15,8 +15,8 @@ class CreateRatesTable extends Migration
     {
         Schema::create('rates', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('client')->nullable();
-            $table->integer('driver')->nullable();
+            $table->float('client', 2, 1)->nullable();
+            $table->float('driver', 2, 1)->nullable();
             $table->text('client_comment')->nullable();
             $table->text('driver_comment')->nullable();
             $table->unsignedInteger('trip_id')->unique();
